#VMS-306 fixed: if camera recording switched on HLS stream buffering never stopped
#VMS-306 low quality stream is now always buffered fore HLS if camera recording switched on. High quality stream is buffered only for single stream camera and low resolution
#VMS-306 removed debug code
(#VMS-313) - add audit for system name change, system merge, general and email settings change
(#VMS-313) - add cameras tab
(#VMS-313) - add option to enable audit trail
(#VMS-313) - alternate colors with grouping by session
(#VMS-313) - bug fixes
(#VMS-313) - change json to ubjson format
(#VMS-313) - color customization
(#VMS-313) - implement audit logging for removing reousrces, database restore and business rule reset
(#VMS-313) - implement bRuleUpdate event type
(#VMS-313) - implement exportVideo event type
(#VMS-313) - implement merge system audit record type
(#VMS-313) - implement user agent for login type
(#VMS-313) - minor changes
(#VMS-313) - move to main menu and sizeHint optimization
(#VMS-313) - painting
(#VMS-313) - polish column width
(#VMS-313) - strict play range for open video action
(#VMS-313) - text filter in progress
(#VMS-313) - tooltip description
(#VMS-313) - update selection model
(#VMS-366) - Add tooltips to the Recording statistics dialog
(#VMS-367) - Sometimes Mac address disapppears
(#VMSVMS-365) - Need to display cameras that record low-quality only too
(Task #6054) - Add table for audit monitor
(Task #6055) - Add audit related calls in mediaServer
(Task #6065) - Add audit monitor model and dialog
(VMS-313) - Add audit log
(VMS-313) - add checkboxes, html fix a bit
(VMS-313) - add unauthorized logins
(VMS-313) - change detail column list either one or many sessions are selected
(VMS-313) - colors customization
(VMS-313) - more button actions
(VMS-313) - painting
Add utility script for generating root pro file.
Added tag release_2.3.2.9167 for changeset d1ba91e25b02
Fixes NXTOOL-67: Can't run installed application.
Fixes NXTOOL-67: Can't run installed application. Added ..\qml\QtQuick\Controls\ComboBox.qml
Fixes NXTOOL-67: Can't run installed application. Now fixing merge issues
IO monitor minor
Merge with dev_2.3.2
Merge with dev_2.3.2_gui
Merge with dev_2.3.2_gui
NXTOOL-61.Enter password button visibility fix
Merge with dev_2.4.0
Merge with dev_2.4.audit
Merge with prod_2.3.2
Merge with release_2.3.2
Merge with x_VMS-306_hls_stream_buffering
NXTOOL-55: Create date enter text control
NXTOOL-55: fixed unchecking 'set current time' checkbox
NXTOOL-55: review fixes
NXTOOL-55: set oacity to 0.5 when dateedit unavailable
NXTOOL-56: Time enter control
NXTOOL-20: Don't show 'Edit all' button if there is nothing to edit
NXTOOL-56: Time enter control, backspace behaviur fixes
NXTOOL-61. Change message 'different networks' to 'Server is unavailable'
NXTOOL-61. Fixed bug with 'enter password' button visibility
NXTOOL-61. Show 'different netwrks' for the servers from ther anoter network
NXTOOL-67/NXTOOL-70: Fixing merge issues #2
NXTOOL-67/NXTOOL-70: Removing qml pdb and debug files from release directory
NXTOOL-71: Nx Tool. Progress bar. Add message, if operation takes a long time
NXTOOL. Show timezone tooltip under unanabled pickers
VMS-323 ec2 protocol version increment
add missed files
fix possible startup crash in case if admin user does not exist yet
ios client version bump
linux compile errors and warnings (after audit was merged in)
refs #6077 coding style fixes
refs #6077 fixed race between sending response to establish transaction connection request and sending transactions
refs #6077 minor code style fixes

<==== Mercurial data ====>
Branch: prod_2.4.0

+++ common/src/utils/media/media_stream_cache.h
@@ -112,7 +112,7 @@ public:
     void unblockData( int blockingID );
 
     //!Time (millis) from last usage of this object
-    size_t inactivityPeriod() const;
+    qint64 inactivityPeriod() const;
 
     struct MediaPacketContext
     {
